[feat]: Per-pixel training and per-image val
Remove previous step() function and integrate all training and
validation computation inside train() fn. This enables to perform
per-pixel training and per-image validation metrics computation. The
purpose of this change is to be able to compute other test metrics such
aas SSIM and LPIPS which take pairs of g.t. and rendered images as their
input.