Renderer changes from dev branch ready to be merged (#62) (#63)
* [feat]: Improve `get_rays` fn

In relation to #30, this function now is able to handle a richer format
of inputs, as long as they have the shape (..., 3, 4) or (..., 4, 4).

* [factor]: Remove old `rendering` module code

Import new `Renderer` ADT and remove code associated to the now
deprecated `rendering` module. This is in relation to #43.

* [factor]: Implement `eval` method for `Renderer`

Method was already implemented. Update method call in runfile by sending
the model explicitly as a parameter.

* [factor]: Import `Renderer` and instantiate it

This commit resolves #58.

* [factor]: Call renderer in train loop and val step

This commit solves #59

* [factor]: Create fn to update renderer chunksize

This commit follows #43 requirements.

* [factor]: Use set_chunksize method

This commit resolves #60